+++ src/screens/dashboard.js
@@ -1,16 +1,13 @@
-import React from 'react';
 import { createStackNavigator } from 'react-navigation';
-import { HamburgerIcon } from '../shared/UI/icons';
 import Dashboard from '../containers/dashboard/main';
 import headerOptions from './header';
 
 const DashboardStack = createStackNavigator({
   main: {
     screen: Dashboard,
-    navigationOptions: ({ navigation }) => ({
+    navigationOptions: {
       title: 'Dashboard', // DICTIONARY
-      headerLeft: <HamburgerIcon onPress={() => navigation.toggleDrawer()} />,
-    }),
+    },
   },
 },
 {