+++ src/screens/settings.js
@@ -1,16 +1,13 @@
-import React from 'react';
 import { createStackNavigator } from 'react-navigation';
-import { HamburgerIcon } from '../shared/UI/icons';
 import SettingsMenu from '../containers/settings/menu';
 import headerOptions from './header';
 
 const SettingsStack = createStackNavigator({
   menu: {
     screen: SettingsMenu,
-    navigationOptions: ({ navigation }) => ({
+    navigationOptions: {
       title: 'Settings', // DICTIONARY
-      headerLeft: <HamburgerIcon onPress={() => navigation.toggleDrawer()} />,
-    }),
+    },
   },
 },
 {