+++ src/screens/work.js
@@ -1,16 +1,13 @@
-import React from 'react';
 import { createStackNavigator } from 'react-navigation';
-import { HamburgerIcon } from '../shared/UI/icons';
 import WorkPlayer from '../containers/work/player';
 import headerOptions from './header';
 
 const WorkStack = createStackNavigator({
   work: {
     screen: WorkPlayer,
-    navigationOptions: ({ navigation }) => ({
+    navigationOptions: {
       title: 'Work', // DICTIONARY
-      headerLeft: <HamburgerIcon onPress={() => navigation.toggleDrawer()} />,
-    }),
+    },
   },
 },
 {