+++ src/shared/UI/icons/hamburger.js
@@ -2,6 +2,11 @@ import React from 'react';
 import { StyleSheet } from 'react-native';
 import { Icon } from 'react-native-elements';
 
+/*
+  * NOTE: this icon is deprecated since v1.1.0,
+  * drawer navigator was changed to a tab navigator
+*/
+
 const styles = StyleSheet.create({
   icon: {
     margin: 13,